fix sync

+++ orchestrator/storage/storage.go
@@ -497,8 +497,8 @@ var DefaultChainSettingsMap = map[string]interface{}{
 	"static_rpc_only":                     false,
 	"min_batch_fee_hls":                   0.1,
 	"min_tx_fee_hls":                      0.1,
-	"oracle_eth_default_blocks_to_search": uint64(2000),
-	"gas_limit":                           5000000,
+	"oracle_eth_default_blocks_to_search": float64(2000),
+	"gas_limit":                           float64(5000000),
 }
 
 func GetChainSettings(chainId uint64) (map[string]interface{}, error) {
